// src/shared/constants/enums.ts
export enum SessionStatus {
  Pending = 'pending',
  Confirmed = 'confirmed',
  InProgress = 'inProgress',
  Completed = 'completed',
  Cancelled = 'cancelled',
  Rejected = 'rejected',
  Declined = 'declined',
  NewRequest = 'newRequest',
  Failed = 'failed',
}

export enum LiveStatus {
  Upcoming = 'upcoming',
  Ready = 'ready',
  Started = 'started',
  Ending = 'ending',
  Completed = 'completed',
  Failed = 'failed',
} 